+++ netdef/Controllers/InternalController.py
@@ -10,6 +10,7 @@ class InternalController(BaseController.BaseController):
         self.logger = logging.getLogger(self.name)
         self.logger.info("init")
         self.send_events = self.shared.config.config(self.name, "send_events", 0)
+        self.send_init_event = self.shared.config.config(self.name, "send_init_event", 0)
 
     def run(self):
         "Main loop. Will exit when receiving interrupt signal"
@@ -22,9 +23,16 @@ class InternalController(BaseController.BaseController):
     def handle_add_source(self, incoming):
         self.logger.debug("'Add source' event for %s", incoming.key)
         self.add_source(incoming.key, incoming)
+        
         if not self.send_events:
             incoming.status_code = StatusCode.INITIAL
             incoming.get = {}
+
+        if self.send_init_event:
+            incoming.status_code = StatusCode.INITIAL
+            incoming.get = {}
+            self.send_outgoing(incoming)
+
 
     def handle_write_source(self, incoming, value, source_time):
         # vi gjør ikke så mye annet enn å tidsstemple nye verdier